LayoutTests/media/modern-media-controls/media-controller/media-controller-fullscreen-change-expected.txt

Testing the MediaController behavior when entering and leaving fullscreen.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Media entered fullscreen
PASS mediaController.layoutTraits is LayoutTraits.macOS | LayoutTraits.Fullscreen
PASS mediaController.controls instanceof MacOSFullscreenMediaControls is true

Media exited fullscreen
PASS mediaController.layoutTraits is LayoutTraits.macOS | LayoutTraits.Compact
PASS mediaController.controls instanceof MacOSInlineMediaControls is true

PASS successfullyParsed is true

TEST COMPLETE
